現状です。今はkogane_waitに色々スクリプトおいてる
