+++ Assets/Animation/Kogane_AttackMotion.anim
@@ -40071,6 +40071,13 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
+  - time: 1.2083334
+    functionName: TatakitukeActive
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 1.5
     functionName: ActiveFalse
     data: 
@@ -40080,6 +40087,13 @@ AnimationClip:
     messageOptions: 0
   - time: 1.8333334
     functionName: CanMoveTrue
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 2
+    functionName: EffectDestroy
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
