+++ Assets/Animation/NewMotiAnimation.anim
@@ -7935,13 +7935,6 @@ AnimationClip:
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
   m_Events:
-  - time: 0
-    functionName: EffectActive
-    data: 
-    objectReferenceParameter: {fileID: 0}
-    floatParameter: 0
-    intParameter: 0
-    messageOptions: 0
   - time: 0.083333336
     functionName: MotiSoundOn
     data: 
@@ -7949,22 +7942,8 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
-  - time: 1.2083334
-    functionName: HurioroshiActive
-    data: 
-    objectReferenceParameter: {fileID: 0}
-    floatParameter: 0
-    intParameter: 0
-    messageOptions: 0
   - time: 2.0416667
     functionName: ResetE
-    data: 
-    objectReferenceParameter: {fileID: 0}
-    floatParameter: 0
-    intParameter: 0
-    messageOptions: 0
-  - time: 2.0416667
-    functionName: EffectDestroy
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
